Run server in background
