+++ src/esp/bindings_js/tests/server.js
@@ -1,7 +1,6 @@
 import http from "http";
 import finalhandler from "finalhandler";
 import serveStatic from "serve-static";
-import puppeteer from "puppeteer";
 
 export async function createServer() {
   const serve = serveStatic("./");
@@ -17,46 +16,19 @@ export async function createServer() {
           server.close(startServer);
         }
       });
-      server.listen(4004, err => (err ? reject(err) : resolve()));
+      server.listen(4004, "localhost", err => {
+        if (err) {
+          reject("Failed to listen on port 4004");
+        } else {
+          const address = server.address();
+          console.log(`Listening on http://${address.address}:${address.port}`);
+          resolve();
+        }
+      });
     };
     startServer();
   });
   return server;
 }
 
-export async function getBrowserAndPage(url) {
-  const browser = await puppeteer.launch({
-    args: ["--disable-lcd-text"],
-    defaultViewport: { width: 1920, height: 1080 }
-  });
-
-  const page = await browser.newPage();
-
-  await page.goto(url, { waitUntil: "load" });
-
-  return { browser, page };
-}
-
-export async function getServerAndURL(path) {
-  const server = await createServer();
-  const address = server.address();
-  const port = address.port;
-  const url = `http://localhost:${port}/${path}`;
-  return {
-    server,
-    url
-  };
-}
-
-export async function closeBrowserAndServer(browser, server) {
-  browser.close();
-  await new Promise(resolve => server.close(resolve));
-}
-
-export async function generateImage(path) {
-  const { server, url } = await getServerAndURL(path);
-  const { browser, page } = await getBrowserAndPage(url);
-  const screenshot = await page.screenshot();
-  closeBrowserAndServer(browser, server);
-  return screenshot;
-}
+createServer();
