+++ src/esp/bindings_js/tests/viewer_demo.visual.test.js
@@ -1,15 +1,14 @@
 import {
-  getServerAndURL,
+  getURL,
   getBrowserAndPage,
-  closeBrowserAndServer
-} from "./server.js";
+} from "./test_utils.js";
 import { toMatchImageSnapshot } from "jest-image-snapshot";
 
 expect.extend({ toMatchImageSnapshot });
 
 test("viewer rendering should match the snapshot", async () => {
   jest.setTimeout(120000);
-  const { server, url } = await getServerAndURL(
+  const url = getURL(
     "build_js/esp/bindings_js/viewer.html?scene=skokloster-castle.glb&useDefaultEpisode=true"
     );
   const { browser, page } = await getBrowserAndPage(url);
@@ -20,7 +19,6 @@ test("viewer rendering should match the snapshot", async () => {
     'document.querySelector("#status").style.color === "white"'
   );
   const screenshot = await page.screenshot();
-
-  closeBrowserAndServer(browser, server);
+  browser.close();
   expect(screenshot).toMatchImageSnapshot();
 });
